# Add benchmarks for triangulation