--- a/napari/benchmarks/benchmark_shapes_layer.py
+++ b/napari/benchmarks/benchmark_shapes_layer.py
@@ -2,17 +2,24 @@
 # https://asv.readthedocs.io/en/latest/writing_benchmarks.html
 # or the napari documentation on benchmarking
 # https://github.com/napari/napari/blob/main/docs/BENCHMARKS.md
+import itertools
 import os
+from contextlib import suppress
+from functools import cache, wraps
 
 import numpy as np
 
 from napari.layers import Shapes
+from napari.layers.shapes._shapes_constants import shape_classes
+from napari.settings import get_settings
 from napari.utils._test_utils import read_only_mouse_event
 from napari.utils.interactions import (
     mouse_move_callbacks,
     mouse_press_callbacks,
     mouse_release_callbacks,
 )
+
+from .utils import Skip
 
 
 class Shapes2DSuite:
@@ -210,6 +217,228 @@ class ShapesInteractionSuite:
         mouse_release_callbacks(self.layer, release_event)
 
     time_select_shape.param_names = ['n_shapes']
+
+
+def get_shape_type(func):
+    @wraps(func)
+    def wrap(self, *args):
+        shape_type_pos = self.param_names.index('shape_type')
+        return func(self, args[shape_type_pos])
+
+    return wrap
+
+
+class _ShapeTriangulationBase:
+    def select_backend(self, compiled_triangulation):
+        with suppress(AttributeError):
+            self.prev_settings = (
+                get_settings().experimental.compiled_triangulation
+            )
+            get_settings().experimental.compiled_triangulation = (
+                compiled_triangulation
+            )
+
+        from napari.layers.shapes import _shapes_utils
+
+        self.triangulate = _shapes_utils.triangulate
+        _shapes_utils.triangulate = None
+
+    def revert_backend(self):
+        with suppress(AttributeError):
+            get_settings().experimental.compiled_triangulation = (
+                self.prev_settings
+            )
+
+        from napari.layers.shapes import _shapes_utils
+
+        _shapes_utils.triangulate = self.triangulate
+
+    def teardown(self, *_):
+        self.revert_backend()
+
+    @get_shape_type
+    def time_create_layer(self, shape_type):
+        """Time to create a layer."""
+        Shapes(self.data, shape_type=shape_type)
+
+    @get_shape_type
+    def time_create_shapes(self, shape_type):
+        """Time to create a layer."""
+        cls = shape_classes[shape_type]
+        for data in self.data:
+            cls(data)
+
+
+class ShapeTriangulationNonConvexSuite(_ShapeTriangulationBase):
+    param_names = [
+        'n_shapes',
+        'n_points',
+        'shape_type',
+        'compiled_triangulation',
+    ]
+    params = [
+        (
+            100,
+            5_000,
+        ),
+        (8, 32, 128),
+        ('path', 'polygon'),
+        (True, False),
+    ]
+
+    # the case of 128 points crashes the benchmark on call of PolygonData(vertices=data).triangulate()
+    skip_params = Skip(
+        if_in_pr=lambda n_shapes,
+        n_points,
+        shape_type,
+        compiled_triangulation: n_shapes > 1000 or n_points > 10,
+        always=lambda n_shapes,
+        n_points,
+        shape_type,
+        compiled_triangulation: not compiled_triangulation
+        and n_points > 100
+        and shape_type == 'polygon',
+    )
+
+    def setup(self, n_shapes, n_points, shape_type, compiled_triangulation):
+        self.data = non_convex_cords(n_shapes, n_points)
+        self.select_backend(compiled_triangulation)
+
+
+class ShapeTriangulationConvexSuite(_ShapeTriangulationBase):
+    param_names = [
+        'n_shapes',
+        'n_points',
+        'shape_type',
+        'compiled_triangulation',
+    ]
+    params = [
+        (
+            100,
+            5_000,
+        ),
+        (8, 32, 128),
+        ('path', 'polygon'),
+        (True, False),
+    ]
+
+    # the case of 128 points crashes the benchmark on call of PolygonData(vertices=data).triangulate()
+    skip_params = Skip(
+        if_in_pr=lambda n_shapes,
+        n_points,
+        shape_type,
+        compiled_triangulation: n_shapes > 1000 or n_points > 10,
+        always=lambda n_shapes,
+        n_points,
+        shape_type,
+        compiled_triangulation: not compiled_triangulation
+        and n_points > 100
+        and shape_type == 'polygon',
+    )
+
+    def setup(self, n_shapes, n_points, shape_type, compiled_triangulation):
+        self.data = convex_cords(n_shapes, n_points)
+        self.select_backend(compiled_triangulation)
+
+
+class ShapeTriangulationMixed(_ShapeTriangulationBase):
+    param_names = ['n_shapes', 'shape_type', 'compiled_triangulation']
+    params = [
+        (
+            100,
+            5_000,
+        ),
+        ('path', 'polygon'),
+        (True, False),
+    ]
+
+    # the case of 128 points crashes the benchmark on call of PolygonData(vertices=data).triangulate()
+    skip_params = Skip(
+        if_in_pr=lambda n_shapes, shape_type, compiled_triangulation: n_shapes
+        > 1000,
+    )
+
+    def setup(self, n_shapes, shape_type, compiled_triangulation):
+        part_size = int(n_shapes / 10)
+        self.data = list(
+            itertools.chain(
+                convex_cords(part_size, 4),
+                convex_cords(part_size * 2, 5),
+                convex_cords(part_size * 2, 7),
+                convex_cords(part_size, 60),
+                non_convex_cords(part_size, 10),
+                non_convex_cords(part_size, 60),
+            )
+        )
+        self.select_backend(compiled_triangulation)
+
+
+@cache
+def non_convex_cords(n_shapes=5_000, n_points=32):
+    """
+    Create a set of non-convex coordinates
+
+    Parameters
+    ----------
+    n_shapes : int
+        Number of shapes to create
+    n_points : int
+        Number of virtex of each shape
+    """
+    rng = np.random.default_rng(0)
+    radius = 500 / np.sqrt(n_shapes)
+    center = rng.uniform(0, 1000, (n_shapes, 2))
+    phi = np.linspace(0, 2 * np.pi, n_points)
+    rays = np.stack([np.sin(phi), np.cos(phi)], axis=1)
+    rays = rays.reshape((1, -1, 2))
+    rays = rays * rng.uniform(0.9, 1.1, (n_shapes, n_points, 2))
+    center = center.reshape((-1, 1, 2))
+    return center + radius * rays
+
+
+@cache
+def self_intersecting_cords(n_shapes=5_000, n_points=32):
+    """
+    Create a set of non-convex coordinates
+
+    Parameters
+    ----------
+    n_shapes : int
+        Number of shapes to create
+    n_points : int
+        Number of virtex of each shape
+    """
+    rng = np.random.default_rng(0)
+    radius = 500 / np.sqrt(n_shapes)
+    center = rng.uniform(0, 1000, (n_shapes, 2))
+    phi = np.linspace(0, 2 * np.pi, n_points)
+    rays = np.stack([np.sin(phi), np.cos(phi)], axis=1)
+    rays = rays.reshape((1, -1, 2))
+    rays = rays * rng.uniform(0.9, 1.1, (n_shapes, n_points, 2))
+    center = center.reshape((-1, 1, 2))
+    return center + radius * rays
+
+
+@cache
+def convex_cords(n_shapes=5_000, n_points=32):
+    """
+    Create a set of convex coordinates
+
+    Parameters
+    ----------
+    n_shapes : int
+        Number of shapes to create
+    n_points : int
+        Number of virtex of each shape
+    """
+    rng = np.random.default_rng(0)
+    radius = 500 / np.sqrt(n_shapes)
+    center = rng.uniform(0, 1000, (n_shapes, 2))
+    phi = np.linspace(0, 2 * np.pi, n_points)
+    rays = np.stack([np.sin(phi), np.cos(phi)], axis=1)
+    rays = rays.reshape((1, -1, 2))
+    center = center.reshape((-1, 1, 2))
+    return center + radius * rays
 
 
 if __name__ == '__main__':