# References and relevant issues

# Description

```
| Change   | Before [a8c822cb] <v0.5.2rc0^0>   | After [6313a39d] <benchmarks_triangulation>   |   Ratio | Benchmark (Parameter)                                                         |
|----------|-----------------------------------|-----------------------------------------------|---------|-------------------------------------------------------------------------------|
| -        | 551±9ms                           | 361±3ms                                       |    0.66 | benchmark_shapes_layer.MeshCalculationSuite.time_set_meshes('polygon', False) |
| -        | 18.6±1ms                          | 3.68±0.1ms                                    |    0.2  | benchmark_shapes_layer.MeshCalculationSuite.time_set_meshes('path', False)    |
| -        | 17.8±0.4ms                        | 486±7μs                                       |    0.03 | benchmark_shapes_layer.MeshCalculationSuite.time_set_meshes('path', True)     |
| -        | 543±7ms                           | 5.45±0.1ms                                    |    0.01 | benchmark_shapes_layer.MeshCalculationSuite.time_set_meshes('polygon', True)  |
```