@Czaki 2025-03-13
I need to add a benchmark for polygons with holes.