@Czaki 2025-03-14
Added improvements and polygon with holes test suites.